HiitのGageWidgetをgeneralへ移動
